Phase 2.4: split contexts/hooks out of provider files for Fast Refresh (31->27)
